// src/app/models/layout_1.ts
import { Layout } from './layout';

export class Layout1 extends Layout {
    card1BotaoLink: string;
    card1BotaoTexto: string;
    card1Descricao: string;
    card1Imagem: string;
    card1Titulo: string;
    card2BotaoLink: string;
    card2BotaoTexto: string;
    card2Descricao: string;
    card2Imagem: string;
    card2Titulo: string;
    ilustracao: string;
}